AgentHost - 💄 extract changeset service interface
